#67: rename getCards to getMaterialsData and remove ListInfo
